container/runner/requirements.txt
anthropic~=0.40
google-generativeai~=0.8
openai~=1.55
aiohttp~=3.9
